requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
click==8.1.7
Deprecated==1.2.14
fastapi==0.110.1
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
protobuf==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
setuptools==69.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.0
typing_extensions==4.10.0
uvicorn==0.29.0
wrapt==1.16.0
zipp==3.18.1
